requirements.txt
fastapi==0.112.1
nats-py==2.8.0
pydantic-settings==2.4.0
uvicorn[standard]==0.30.6
